+++ GFramework.Core.Tests/Architectures/ArchitectureContextTests.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
-using GFramework.Core.Abstractions.Architecture;
+using GFramework.Core.Abstractions.Architectures;
 using GFramework.Core.Abstractions.Command;
 using GFramework.Core.Abstractions.Enums;
 using GFramework.Core.Abstractions.Environment;
