+++ GFramework.Core.Tests/Coroutine/MediatorCoroutineExtensionsTests.cs
@@ -11,13 +11,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using GFramework.Core.Abstractions.Architecture;
+using GFramework.Core.Abstractions.Architectures;
 using GFramework.Core.Abstractions.Coroutine;
 using GFramework.Core.Abstractions.Rule;
 using GFramework.Core.Coroutine.Extensions;
 using Mediator;
 using Moq;
-using NUnit.Framework;
 
 namespace GFramework.Core.Tests.Coroutine;
