+++ GFramework.Core.Tests/Coroutine/QueryCoroutineExtensionsTests.cs
@@ -1,10 +1,9 @@
-using GFramework.Core.Abstractions.Architecture;
+using GFramework.Core.Abstractions.Architectures;
 using GFramework.Core.Abstractions.Coroutine;
 using GFramework.Core.Abstractions.Query;
 using GFramework.Core.Abstractions.Rule;
 using GFramework.Core.Coroutine.Extensions;
 using Moq;
-using NUnit.Framework;
 
 namespace GFramework.Core.Tests.Coroutine;
