+++ GFramework.Core.Tests/Mediator/MediatorArchitectureIntegrationTests.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics;
 using System.Reflection;
-using GFramework.Core.Abstractions.Architecture;
+using GFramework.Core.Abstractions.Architectures;
 using GFramework.Core.Architectures;
 using GFramework.Core.Command;
 using GFramework.Core.Ioc;
