+++ GFramework.Core.Tests/Rule/ContextAwareServiceExtensionsTests.cs
@@ -1,4 +1,4 @@
-using GFramework.Core.Abstractions.Architecture;
+using GFramework.Core.Abstractions.Architectures;
 using GFramework.Core.Abstractions.Enums;
 using GFramework.Core.Abstractions.Model;
 using GFramework.Core.Abstractions.Rule;
