+++ GFramework.Core/Architectures/ArchitectureContext.cs
@@ -1,5 +1,5 @@
 using System.Collections.Concurrent;
-using GFramework.Core.Abstractions.Architecture;
+using GFramework.Core.Abstractions.Architectures;
 using GFramework.Core.Abstractions.Command;
 using GFramework.Core.Abstractions.Environment;
 using GFramework.Core.Abstractions.Events;
