+++ GFramework.Core/Architectures/GameContext.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Concurrent;
-using GFramework.Core.Abstractions.Architecture;
+using GFramework.Core.Abstractions.Architectures;
 
 namespace GFramework.Core.Architectures;
